I completed the game
I haven't been uploading this game because well, I never needed to and never changed systems during the time of development. Here I replicated the player's moment with a Rigidbody, and tried to combine the camera system as close to the original as possible. The camera gimmick took the most time but hey, it's done.